Moved `NumberHelper` into `org.apache.tinkerpop.gremlin.util`.
