+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/process/traversal/step/map/MaxLocalStep.java
@@ -26,7 +26,7 @@ import java.util.Collections;
 import java.util.Iterator;
 import java.util.Set;
 
-import static org.apache.tinkerpop.gremlin.process.traversal.NumberHelper.max;
+import static org.apache.tinkerpop.gremlin.util.NumberHelper.max;
 
 /**
  * @author Marko A. Rodriguez (http://markorodriguez.com)
